refactor: 스토리북을 좀 더 나은 구조로 폴더링
