+++ src/components/Button/Button.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import Button from "./index";
 
 export default {
-  title: "Example/Button",
+  title: "Components/Button",
   component: Button,
   argTypes: {},
 };
