+++ src/components/Header/Header.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import Header from "./index";
 
 export default {
-  title: "Example/Header",
+  title: "Components/Header",
   component: Header,
   argTypes: {},
 };
