--- src/components/common/Button/style.css
@@ -1,16 +1,0 @@
-.button {
-  outline: none;
-  border: 0;
-  color: #04c09e;
-  font-size: 1.1rem;
-  cursor: pointer;
-  border-radius: 10px;
-  padding: 0.5rem 0.6rem 0.3rem;
-  background-color: rgba(0, 0, 0, 0);
-
-  display: flex;
-}
-
-.button:hover {
-  background-color: #e7e7e7;
-}
