--- src/components/common/CardTypeRadio/style.css
@@ -1,40 +1,0 @@
-.card-type-radio {
-  position: relative;
-  display: inline-flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.card-type-radio__logo {
-  position: relative;
-  width: 3rem;
-  height: 3rem;
-  border-radius: 50%;
-  overflow: hidden;
-}
-
-.card-type-radio > p {
-  margin: 0.5rem 0.6rem 0;
-  font-size: 0.9rem;
-}
-
-.card-type-radio > input {
-  position: absolute;
-  opacity: 0;
-}
-
-.card-type-radio > input:checked ~ .card-type-radio__logo::after {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: radial-gradient(
-    50% 50% at 50% 50%,
-    rgba(0, 0, 0, 0) 0%,
-    rgba(0, 0, 0, 0) 70%,
-    rgba(0, 0, 0, 0.15) 80%,
-    rgba(0, 0, 0, 0.25) 100%
-  );
-}
