--- src/components/common/Header/style.css
@@ -1,39 +1,0 @@
-.header {
-  display: flex;
-  align-items: center;
-}
-
-.header > button {
-  position: relative;
-  display: flex;
-  align-items: center;
-  height: 2rem;
-  width: 2rem;
-  border-radius: 50%;
-  border: none;
-  outline: none;
-  cursor: pointer;
-  background-color: rgba(0, 0, 0, 0);
-}
-
-.header > button::after {
-  position: absolute;
-  left: 35%;
-  content: "";
-  display: inline-block;
-  height: 0.6rem;
-  width: 0.6rem;
-  border-left: 2px solid #525252;
-  border-bottom: 2px solid #525252;
-  transform: rotate(45deg);
-}
-
-.header > button:hover {
-  background-color: #e7e7e7;
-}
-
-.header > h2 {
-  font-size: 1.1rem;
-  line-height: 2.5;
-  margin-left: 0.5rem;
-}
