--- src/components/common/Modal/style.css
@@ -1,37 +1,0 @@
-.modal {
-  display: flex;
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.5);
-  transition: opacity 0.25s ease;
-  z-index: 2;
-
-  justify-content: center;
-  align-items: flex-end;
-}
-
-.modal__inner {
-  min-height: 30vh;
-  width: min(100vw, 500px);
-  background-color: #fff;
-  border-top-left-radius: 15px;
-  border-top-right-radius: 15px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  animation: appear 0.2s ease-in-out;
-}
-
-@keyframes appear {
-  0% {
-    transform: translateY(100%);
-  }
-
-  100% {
-    transform: translateY(0);
-  }
-}
