--- src/components/common/constants/card.js
@@ -1,44 +1,0 @@
-export const CARD = {
-  POCO: {
-    name: "포코 카드",
-    color: "red",
-  },
-  JUN: {
-    name: "준 카드",
-    color: "blue",
-  },
-  PARK: {
-    name: "공원 카드",
-    color: "green",
-  },
-  BRAN: {
-    name: "브랜 카드",
-    color: "pink",
-  },
-  LOYD: {
-    name: "로이드 카드",
-    color: "aqua",
-  },
-  DOBI: {
-    name: "도비 카드",
-    color: "light-pink",
-  },
-  COLIN: {
-    name: "콜린 카드",
-    color: "orange",
-  },
-  SUN: {
-    name: "썬 카드",
-    color: "yellow",
-  },
-  UNKNOWN: {
-    name: "",
-    color: "gray",
-  },
-};
-
-export const CARD_SIZE = {
-  SMALL: "small",
-  MEDIUM: "medium",
-  LARGE: "large",
-};
