--- src/components/common/style/background.css
@@ -1,35 +1,0 @@
-.bg-gray {
-  background-color: #d2d2d2;
-}
-
-.bg-red {
-  background-color: #e24141;
-}
-
-.bg-blue {
-  background-color: #547ce4;
-}
-
-.bg-green {
-  background-color: #73bc6d;
-}
-
-.bg-pink {
-  background-color: #de59b9;
-}
-
-.bg-aqua {
-  background-color: #04c09e;
-}
-
-.bg-light-pink {
-  background-color: #e76e9a;
-}
-
-.bg-orange {
-  background-color: #f37d3b;
-}
-
-.bg-yellow {
-  background-color: #fbcd58;
-}
