+++ src/components/style/background.css
