+++ src/pages/CompleteCardAddition.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import { CARD_SIZE } from "./common/constants/card";
+import { CARD_SIZE } from "../constants";
 import PropTypes from "prop-types";
-import { Card, Button } from "./common";
+import { Card, Button } from "../components";
 import { CARD_DESCRIPTION, FORMAT_CHAR, URL } from "../constants";
 import { Redirect } from "react-router";
